requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
alembic>=1.13.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Database drivers
# For PostgreSQL/Xata
psycopg2-binary>=2.9.9

# For HTTP requests
requests>=2.31.0 